requirements.txt
certifi==2019.11.28
coverage==5.1
Django==2.2.11
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-filter==2.2.0
django-nine==0.2.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
